moved files to entity-architecture
