+++ _helpers/account-verification-email.js
@@ -1,6 +1,6 @@
 require('dotenv').config();
 
-const User = require('../models/User');
+const User = require('../user/user.model');
 
 const nodemailer = require('nodemailer');
 
@@ -13,7 +13,7 @@ const emailTransporter = nodemailer.createTransport(
         apiKey: process.env.SENDGRID_API_KEY
     }));
 
-module.exports = async (userEmail) => {
+module.exports = async (userId) => {
     try {
         crypto.randomBytes(32, async (err, buffer) => {
             try {
@@ -22,21 +22,20 @@ module.exports = async (userEmail) => {
                 }
                 const token = buffer.toString('hex');
 
-                const user = await User.findOne({ email: userEmail });
-                if (!user) {
-                    throw Error('User does not exist');
+                const user = await User.findById(userId);
+                if (user.active === true) {
+                    throw Error('User is already verified.');
                 }
-
-                user.passwordReset = { token: token, expires: Date.now() + 3600000 };
+                user.verification = { token: token, expires: Date.now() + 3600000 };
                 const savedUser = await user.save();
 
                 sentEmail = await emailTransporter.sendMail({
                     to: savedUser.email,
                     from: 'restapi201@gmail.com',
-                    subject: 'Cardgames - Password Reset',
-                    html: `<h1>To reset your password, please click on link below</h1>
+                    subject: 'Cardgames Registration',
+                    html: `<h1>To complete registration, please click on link below</h1>
                     <br>
-                    http://localhost:`+ process.env.PORT + `/user/` + savedUser._id + `/password_reset/` + savedUser.passwordReset.token
+                    <a href="http://localhost:`+ process.env.PORT + `/user/` + savedUser._id + `/verify/` + savedUser.verification.token + `">VERIFY ACCOUNT</a>`
                 });
             } catch (err) {
                 throw err;
