+++ game/use-cases/game-definition/create.js
@@ -1,19 +1,17 @@
-const Game = require('../../../models/Game');
-const User = require('../../../models/User');
+const Game = require('../../game.model');
 
 module.exports = async (req, res, next) => {
     const userId = req.userId;
     const gameType = req.params.gameType;
 
     try {
-        const user = await User.findById(userId);
         const game = new Game({
             status: 'waiting',
             gameType: gameType,
             players: [{
                 order: game.players.length,
                 userType: 'player',
-                userId: user._id,
+                userId: userId,
                 pointsTotal: 0,
                 pointsCurrentGame: 0
             }]
