+++ game/use-cases/game-definition/leave.js
@@ -1,5 +1,4 @@
-const Game = require('../../../models/Game');
-const Player = require('../../../models/User');
+const Game = require('../../game.model');
 
 module.exports = (req, res, next) => {
     const userId = req.userId;
