+++ game/use-cases/game-specific/callbreak/start.js
@@ -1,29 +1,30 @@
-const Game = require('../../../models/Game');
-const User = require('../../../models/User');
-const customError = require('../../../helpers/custom-error');
-const Deck = require('../../../entities/Deck');
+const Game = require('../../../game.model');
+
+const Deck = require('../../../../entities/Deck');
+
+const customError = require('../../../../_helpers/custom-error');
 
 module.exports = async (req, res, next) => {
     const userId = req.userId;
     const gameType = req.params.gameType;
     const gameId = req.params.gameId;
+
     const deck = new Deck();
     const dealtCardsObject = deck.dealCards(13, 4); //dealing 13 cards to 4 players
     const remainingCards = dealtCardsObject.remaining;
 
     try {
-        const user = await User.findById(userId);
         const game = await Game.findById(gameId);
 
         if (!game) {
             throw customError('The game does not exist.', 404, null);
         }
 
-        const isValidPlayer = game.players.map(x => x.userId).includes(user._id);
+        const isValidPlayer = game.players.map(x => x.userId).includes(userId);
 
         if (isValidPlayer && game.status == 'waiting') {
             while (game.players.length < 4) {
-                game.players.push({ userType: 'bot', userId: user._id });
+                game.players.push({ userType: 'bot', userId: userId });
             }
 
             for (let i = 0; i < 4; i++) {
