--- routes/game-specific/callbreak.js
@@ -1,7 +1,0 @@
-const router = require('express').Router();
-
-const startCallbreak = require('../../controllers/game/callbreak/start');
-
-router.get('/:gameId/start', startCallbreak); // success method 200
-
-module.exports = router;
